add validation id param

+++ desafio04/routes/products.js
@@ -3,6 +3,8 @@ const productsRouter = express.Router();
 
 const controller = require('../controller');
 
+const error = { error: 'Producto no encontrado' }
+
 productsRouter.get("/", (req, res) => {
     const products = controller.getAll();
     res.send(products);
@@ -11,13 +13,24 @@ productsRouter.get("/", (req, res) => {
   productsRouter.get("/:id", (req, res) => {
     const productId = parseInt(req?.params?.id);
     const product = controller.getById(productId);
-    res.send(product);
+
+    if(product){
+      res.send(product);
+    } else {
+      res.send(error)
+    }
   });
 
   productsRouter.delete("/:id", (req, res) => {
     const productId = parseInt(req?.params?.id);
-    const product = controller.deleteById(productId);
-    res.send(product);
+    const product = controller.getById(productId);
+
+    if(product) {
+      controller.deleteById(productId);
+      res.send(`Producto ${productId} eliminado`);
+    } else {
+      res.send(error);
+    }
   });
 
   productsRouter.post("/", (req, res) => {
@@ -27,8 +40,14 @@ productsRouter.get("/", (req, res) => {
 
   productsRouter.put("/:id", (req, res) => {
     const productId = parseInt(req?.params?.id);
-    const product = controller.update(productId, req.body);
-    res.status(201).send(product);
+    const product = controller.getById(productId);
+
+    if(product){
+      controller.update(productId, req.body);
+      res.status(201).send(product);
+    } else {
+      res.send(error);
+    }
   });
 
 module.exports = productsRouter;
